pyproject.toml
[project]
name = "release-manager"
version = "0.1.0"
description = "CrewAI project that walks through the release checklist using Notion, Slack, and Jira."
authors = [{name = "ManagerOps-Factory"}]
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.28.0",
    "typer>=0.12.3",
    "rich>=13.7.1",
    "pyyaml>=6.0.1"
]

[project.optional-dependencies]
dev = ["ruff"]
